Check file is not empty

+++ backup_no_encryption.sh
@@ -8,6 +8,8 @@ TOKEN=${NETRC[1]}
 
 curl "https://api.pinboard.in/v1/posts/all?format=json&auth_token=$USERNAME:$TOKEN" | python -mjson.tool > /tmp/${JSONFILE}
 
-mv /tmp/${JSONFILE} ~/Documents/Pinboard/
-
-ln -sf ~/Documents/Pinboard/${JSONFILE} ~/Documents/Pinboard/Pinboard.json
+if [ -s "/tmp/${JSONFILE}" ]
+then
+  mv /tmp/${JSONFILE} ~/Documents/Pinboard/
+  ln -sf ~/Documents/Pinboard/${JSONFILE} ~/Documents/Pinboard/Pinboard.json
+fi
